Move iterable-partitioning utility methods to IterablePartitioner
There's some weirdness here with passing loggers around, but to
minimise the footprint of this change, I preserved it.
